+++ stroom-pipeline/src/test/java/stroom/pipeline/refdata/store/offheapstore/databases/AbstractLmdbDbTest.java
@@ -19,10 +19,12 @@ package stroom.pipeline.refdata.store.offheapstore.databases;
 
 import stroom.lmdb.LmdbEnv;
 import stroom.lmdb.LmdbEnvFactory;
+import stroom.lmdb.LmdbLibraryConfig;
 import stroom.test.common.util.test.StroomUnitTest;
 import stroom.util.io.ByteSize;
 import stroom.util.io.FileUtil;
 import stroom.util.io.PathCreator;
+import stroom.util.io.TempDirProvider;
 import stroom.util.shared.ModelStringUtil;
 
 import org.junit.jupiter.api.AfterEach;
@@ -56,8 +58,9 @@ public abstract class AbstractLmdbDbTest extends StroomUnitTest {
                 Arrays.toString(envFlags));
 
         final PathCreator pathCreator = new PathCreator(() -> dbDir, () -> dbDir);
+        final TempDirProvider tempDirProvider = () -> dbDir;
 
-        lmdbEnv = new LmdbEnvFactory(pathCreator)
+        lmdbEnv = new LmdbEnvFactory(pathCreator, tempDirProvider, new LmdbLibraryConfig())
                 .builder(dbDir)
                 .withMapSize(getMaxSizeBytes())
                 .withMaxDbCount(10)