+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Issue **#2478** : Create a single place in config for the LMDB library path and extraction dir.
+
+* Issue **#2478** : Delete old LMDB library binaries on boot.
+
 * Issue **#2497** : Fixed issue with `OrphanFileFinder` incorrectly identifying some dirs as being empty.
 
 * Change `/api/refData/v1/purgeByAge/{purgeAge}`, `/api/refData/v1/purgeByStream/{refStreamId}` and `/api/refData/v1/clearBufferPool` to act on all nodes unless the `nodeName` query param is provided.